moved config to global configuration and improved documentation
