--- src/main/resources/org/jvnet/hudson/plugins/thinbackup/ThinBackupMgmtLink/backupsettings.jelly
@@ -1,135 +1,0 @@
-<!--
-  Copyright (C) 2011  Matthias Steinkogler, Thomas Fürer
-
-  This program is free software: you can redistribute it and/or modify
-  it under the terms of the GNU General Public License as published by
-  the Free Software Foundation, either version 3 of the License, or
-  (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU General Public License for more details
-
-  You should have received a copy of the GNU General Public License
-  along with this program.  If not, see http://www.gnu.org/licenses.
-
- -->
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
-         xmlns:f="/lib/form">
-    <l:layout permission="${app.ADMINISTER}" norefresh="true" title="">
-        <st:include it="${app}" page="sidepanel.jelly"/>
-        <l:main-panel>
-            <h1>
-                ${%thin_backup_configuration}
-            </h1>
-
-            <f:form method="post" action="saveSettings" name="saveSettings">
-                 <f:section title="${%backup_settings}">
-                    <f:entry title="${%backup_path}" field="backupPath"
-                        help="/plugin/thinBackup/help/help-backupPath.html">
-                        <f:textbox value="${it.configuration.backupPath}" name="backupPath"
-                                   checkMethod="post" checkUrl="'checkBackupPath?value='+escape(this.value)"/>
-<!--
-                                   checkMethod="post" checkUrl="'${rootURL}/plugin/thinBackup/checkBackupPath?value='+escape(this.value)"/>
--->
-                    </f:entry>
-                    
-                    <f:entry title="${%full_backup_schedule}" field="fullBackupSchedule"
-                        help="/plugin/thinBackup/help/help-fullBackupSchedule.html">
-                        <f:textbox value="${it.configuration.fullBackupSchedule}" name="fullBackupSchedule"/>
-                    </f:entry>
-                    
-                    <f:entry title="${%diff_backup_schedule}" field="diffBackupSchedule"
-                        help="/plugin/thinBackup/help/help-diffBackupSchedule.html">
-                        <f:textbox value="${it.configuration.diffBackupSchedule}" name="diffBackupSchedule"/>
-                    </f:entry>
-                    
-                    <f:entry title="${%max_backup_full}" field="nrMaxStoredFull"
-                        help="/plugin/thinBackup/help/help-nrMaxStoredFull.html">
-                        <f:textbox value="${it.configuration.nrMaxStoredFull}" name="nrMaxStoredFull" />
-                    </f:entry>
-                    
-                    <f:entry title="${%exclude_files_regex}" field="excludedFilesRegex"
-                        help="/plugin/thinBackup/help/help-excludedFilesRegex.html">
-                        <f:textbox value="${it.configuration.excludedFilesRegex}" name="excludedFilesRegex"/>
-                    </f:entry>
-                    
-                    <f:optionalBlock title="${%wait_for_idle}"
-                        help="/plugin/thinBackup/help/help-waitForIdle.html"
-                        checked="${it.configuration.waitForIdle}"
-                        name="waitForIdle">
-                        <f:entry title="${%force_quite_minutes}" field="forceQuietModeTimeout"
-	                        help="/plugin/thinBackup/help/help-forceQuietModeTimeout.html">
-	                        <f:textbox value="${it.configuration.forceQuietModeTimeout}" name="forceQuietModeTimeout"/>
-                        </f:entry>
-                    </f:optionalBlock>
-
-                    <f:optionalBlock title="${%backup_results}"
-                        help="/plugin/thinBackup/help/help-backupBuildResults.html"
-                        checked="${it.configuration.backupBuildResults}"
-                        name="backupBuildResults" >
-                        <f:optionalBlock title="${%backup_build_archive}"
-                            help="/plugin/thinBackup/help/help-backupBuildArchive.html"
-                            checked="${it.configuration.backupBuildArchive}"
-                            name="backupBuildArchive" />
-                        <f:optionalBlock title="${%backup_build_keep}"
-                            help="/plugin/thinBackup/help/help-backupBuildToKeepOnly.html"
-                            checked="${it.configuration.backupBuildsToKeepOnly}"
-                            name="backupBuildsToKeepOnly" />
-                    </f:optionalBlock>
-                    
-                    <f:optionalBlock title="${%backup_user_content}"
-                        help="/plugin/thinBackup/help/help-backupUserContent.html"
-                        checked="${it.configuration.backupUserContents}"
-                        name="backupUserContents" />
-
-                    <f:optionalBlock title="${%backup_build_number}"
-                        help="/plugin/thinBackup/help/help-backupNextBuildNumber.html"
-                        checked="${it.configuration.backupNextBuildNumber}"
-                        name="backupNextBuildNumber" />
-
-                    <f:optionalBlock title="${%backup_plugin_archive}"
-                        help="/plugin/thinBackup/help/help-backupPlugins.html"
-                        checked="${it.configuration.backupPluginArchives}"
-                        name="backupPluginArchives" >
-                    </f:optionalBlock>
-
-                    <f:optionalBlock title="${%backup_config_history}"
-                        help="/plugin/thinBackup/help/help-backupConfigHistory.html"
-                        checked="${it.configuration.backupConfigHistory}"
-                        name="backupConfigHistory" >
-                    </f:optionalBlock>
-
-                    <f:optionalBlock title="${%backup_additional_files}"
-                        help="/plugin/thinBackup/help/help-backupAdditionalFiles.html"
-                        checked="${it.configuration.backupAdditionalFiles}"
-                        name="backupAdditionalFiles">
-                        <f:entry title="${%backup_additional_files_regex}" field="backupAdditionalFilesRegex">
-                            <f:textbox value="${it.configuration.backupAdditionalFilesRegex}" name="backupAdditionalFilesRegex"/>
-                        </f:entry>
-                    </f:optionalBlock>
-                    
-                    <f:optionalBlock title="${%clean_differential}"
-                        help="/plugin/thinBackup/help/help-cleanupDiff.html"
-                        checked="${it.configuration.cleanupDiff}"
-                        name="cleanupDiff" />
-                    
-                    <f:optionalBlock title="${%move}"
-                        help="/plugin/thinBackup/help/help-moveOldBackupsToZipFile.html"
-                        checked="${it.configuration.moveOldBackupsToZipFile}"
-                        name="moveOldBackupsToZipFile" />
-                    
-                    <f:optionalBlock title="${%fail_fast}"
-                        help="/plugin/thinBackup/help/help-failFast.html"
-                        checked="${it.configuration.failFast}"
-                        name="failFast" />
-                     <f:entry>
-                         <f:submit value="${%save}"/>
-                     </f:entry>
-                </f:section>
-            </f:form>
-        </l:main-panel>
-    </l:layout>
-</j:jelly>
