+++ src/main/resources/org/jvnet/hudson/plugins/thinbackup/ThinBackupMgmtLink/index.properties
@@ -1,6 +1,6 @@
 backup_manual_tooltip = Click to start the backup now.
 backup_now = Backup now
-backup_settings_tooltip = Click to open plugin settings.
+backup_settings_moved = Settings are now integrated in global configuration.
 restore = Restore
 restore_options_tooltip = Click to open the restore.
 settings = Settings
