+++ src/main/resources/org/jvnet/hudson/plugins/thinbackup/ThinBackupMgmtLink/index_de.properties
@@ -1,6 +1,6 @@
-backup_manual_tooltip = Klicken um das Backup jetzt zu starten.
+backup_manual_tooltip = Klicken, um das Backup jetzt zu starten.
 backup_now = Backup starten
-backup_settings_tooltip = Klicken um zu den Backup Einstellungen zu gelangen.
+backup_settings_moved = Einstellungen sind nun in den globalen Einstellungen integriert.
 restore = Wiederherstellen
-restore_options_tooltip = Klicken um zum Wiederherstellen zu gelangen.
+restore_options_tooltip = Klicken, um zum Wiederherstellen zu gelangen.
 settings = Einstellungen
