+++ src/main/resources/org/jvnet/hudson/plugins/thinbackup/ThinBackupPluginImpl/config.properties
